images/podcast-feed-filter/requirements.txt
flask>=3.0
gunicorn>=23.0
lxml>=5.0
pyyaml>=6.0
requests>=2.32
